+++ src/question/expand/mod.rs
@@ -145,7 +145,7 @@ impl<F: Fn(char) -> Option<char>> ui::Widget for ExpandPrompt<'_, F> {
 
             if self.input.value().is_some() {
                 b.move_cursor(MoveDirection::NextLine(1))?;
-                b.write_styled(&ui::symbols::SMALL_ARROW.cyan())?;
+                b.write_styled(&ui::symbols::current().arrow.cyan())?;
                 b.write_all(b" ")?;
 
                 layout.offset_y += 1;